Move operator enums into query
