--- core/src/main/java/com/redhat/lightblue/client/LogOp.java
@@ -1,16 +1,0 @@
-package com.redhat.lightblue.client;
-
-public enum LogOp {
-    and("$and"),
-    or("$or");
-
-    private String s;
-    
-    private LogOp(String s) {
-        this.s=s;
-    }
-
-    public String toString() {
-        return s;
-    }
-}
