--- core/src/main/java/com/redhat/lightblue/client/NaryOp.java
@@ -1,16 +1,0 @@
-package com.redhat.lightblue.client;
-
-public enum NaryOp {
-    in("$in"),
-    nin("$nin");
-
-    private String s;
-    
-    private NaryOp(String s) {
-        this.s=s;
-    }
-
-    public String toString() {
-        return s;
-    }
-}
